set pipeline latency and add some max-lateness slack
Set the pipeline latency via the processing deadline property to
something that can actually be reached under real system load. Also
add some more slack to the allowed buffer lateness. As the whole
pipeline is running at 30fps, the default 5ms is a bit harsh and
drops frames that are still fine to present.

Signed-off-by: Lucas Stach <l.stach@pengutronix.de>