plugin: remove sync variable
The sync variable allows to disable sync on the GStreamer sink element.
Currently it is set to true and never changed.

With glibc 2.36, the variable conflicts with the definition of the `void
sync()` symbol defined in /usr/include/unistd.h, which is a POSIX
standard function.

Having this variable is doubtful by itself, but conflicting with a POSIX
declaration is bad.

Signed-off-by: Michael Tretter <m.tretter@pengutronix.de>